Fix test hanging issue: unregister Prometheus metrics between tests
Root cause: Prometheus metrics (graph_info, graph_version) were being registered
when creating FlowchartCtrlWidget. When multiple tests ran, the second test
would try to register the same metric names again, causing:
  ValueError: Duplicated timeseries in CollectorRegistry

Fix:
- Added cleanup in flowchart fixture to unregister Prometheus metrics
- Moved broker, dmypy, and flowchart fixtures from test_gui.py to conftest.py
  so they can be shared by test_integration.py
- Fixed test_connect_nodes to use cspad->Roi2D (correct types) instead of laser->Projection

Result:
- All 5 GUI tests now pass when run together: 5 passed in 15.50s
- Tests no longer hang after 3rd test
- Session-scoped backend successfully shared across tests

Note: test_integration.py tests currently fail due to fixture compatibility issues
      with complex_graph_file (uses dill format, not .fc JSON). Will address separately.
